refactor: move server code into server/src
