+++ server/src/db.js
@@ -1,0 +1,3 @@
+const { Pool } = require('pg');
+const pool = new Pool();
+module.exports = pool;
